requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dnspython @ git+https://github.com/rthalley/dnspython.git@6f5ff0c0ea680871818460882c92dbb412c3ff13
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
sniffio==1.3.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
